Stream stability + low-latency fixes; clippy/fmt cleanup
ABR / encoding:
- BitrateVerifier measures real encoder output and escalates a silently-ignored
  in-place downward bitrate change to a background encoder rebuild, so ABR stays
  effective on every backend (libx264 honors in-place; hw-safe, no latency cost).
- Infinite GOP by default (ST_GOP=0 -> infinite), removing the ~1s periodic IDR
  "refresh" spikes; keyframes stay on-demand.

clippy/fmt (warning-free):
- collapse nested if-let into match arms (wl_capture, ext_image_copy, pipewire,
  punched-auth), drop redundant same-type pointer casts (linux_uring, pipewire)
- type aliases for complex types: PipelineSetup, LinuxEncoderChoice, ModifierCache,
  DialogFn; matches! for env flags (io_uring/gso/wake)
- c"" literals in tests; iterator loop; named JoinHandle for detached ctrl-c task
- gate test-only cursor-prediction helpers behind cfg(test)
- rustfmt the crate; bump protocol submodule

+++ src/encode_vaapi.rs
@@ -471,7 +471,7 @@ impl VaapiEncoder {
             for (i, plane) in planes.iter().enumerate() {
                 desc.objects[i].fd = plane.fd.as_raw_fd();
                 desc.objects[i].size = 0;
-                desc.objects[i].format_modifier = plane.modifier as u64;
+                desc.objects[i].format_modifier = plane.modifier;
             }
             desc.nb_layers = 1;
             desc.layers[0].format = _drm_format;
